# [fix](Nereids) string literal coercion of in predicate