@morrySnow 2024-05-23
run buildall